requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==23.1.0
bcrypt==4.1.2
bitarray==2.8.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
cytoolz==0.12.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.5
django-import-export==3.3.7
django-jazzmin==2.6.0
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.1
frozenlist==1.4.0
hexbytes==0.3.1
html5lib==1.1
idna==3.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lru-dict==1.2.0
lxml==5.1.0
MarkupPy==1.14
multidict==6.0.4
mysqlclient==2.2.4
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
parsimonious==0.9.0
pillow==10.2.0
protobuf==4.24.3
pycparser==2.21
pycryptodome==3.19.0
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
pypdf==4.0.2
pypng==0.20220715.0
python-bidi==0.4.2
pyunormalize==15.0.0
pywin32==306
PyYAML==6.0.1
qrcode==7.4.2
razorpay==1.4.1
referencing==0.30.2
regex==2023.8.8
reportlab==4.0.9
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.3
shortuuid==1.0.11
six==1.16.0
SpeechRecognition==3.10.1
sqlparse==0.4.4
svglib==1.5.1
tablib==3.5.0
tinycss2==1.2.1
toolz==0.12.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritools==4.0.2
urllib3==2.0.5
web3==6.10.0
webencodings==0.5.1
websockets==11.0.3
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.2
